Added Decimals to Company Credit

+++ elixir/lib/homework_web/schemas/companies_schema.ex
@@ -9,8 +9,8 @@ defmodule HomeworkWeb.Schemas.CompaniesSchema do
     object :company do
       field(:id, non_null(:id))
       field(:name, :string)
-      field(:credit_line, :integer)
-      field(:available_credit, :integer)
+      field(:credit_line, :float)
+      field(:available_credit, :float)
       field(:inserted_at, :naive_datetime)
       field(:updated_at, :naive_datetime)
     end
@@ -19,7 +19,7 @@ defmodule HomeworkWeb.Schemas.CompaniesSchema do
       @desc "Create a new company"
       field :create_company, :company do
         arg(:name, non_null(:string))
-        arg(:credit_line, non_null(:integer))
+        arg(:credit_line, non_null(:float))
   
         resolve(&CompaniesResolver.create_company/3)
       end
@@ -28,7 +28,7 @@ defmodule HomeworkWeb.Schemas.CompaniesSchema do
       field :update_company, :company do
         arg(:id, non_null(:id))
         arg(:name, non_null(:string))
-        arg(:credit_line, non_null(:integer))
+        arg(:credit_line, non_null(:float))
   
         resolve(&CompaniesResolver.update_company/3)
       end
